refactor: pull collections nav out into shareable component
